requirements.txt
djangorestframework
django-cors-headers
npm install axios react-router-dom recharts